Move SNMP private defines into vrrp_snmp.c/check_snmp.c
The defines for the net-snmp "magic" were in the header files
which were included by other modules. The defines are private to
the c source file, so move the defines into them, to avoid polution
compilation units which included vrrp_snmp.h/check_snmp.h.
